PM-1155 - refactor code
