+++ src/api/admin/admin.module.ts
@@ -1,0 +1,19 @@
+import { Module } from '@nestjs/common';
+import { AdminController } from './admin.controller';
+import { AdminService } from './admin.service';
+import { TaxFormRepository } from '../repository/taxForm.repo';
+import { PaymentMethodRepository } from '../repository/paymentMethod.repo';
+import { WinningsRepository } from '../repository/winnings.repo';
+import { TopcoderModule } from 'src/shared/topcoder/topcoder.module';
+
+@Module({
+  imports: [TopcoderModule],
+  controllers: [AdminController],
+  providers: [
+    AdminService,
+    TaxFormRepository,
+    PaymentMethodRepository,
+    WinningsRepository,
+  ],
+})
+export class AdminModule {}
