+++ src/api/api.module.ts
@@ -1,12 +1,5 @@
 import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { HealthCheckController } from './health-check/healthCheck.controller';
-import { AdminWinningController } from './admin-winning/adminWinning.controller';
-import { AdminWinningService } from './admin-winning/adminWinning.service';
-import { UserWinningController } from './user-winning/userWinning.controller';
-import { WinningController } from './winning/winning.controller';
-import { WinningService } from './winning/winning.service';
-import { WalletController } from './wallet/wallet.controller';
-import { WalletService } from './wallet/wallet.service';
 import { GlobalProvidersModule } from 'src/shared/global/globalProviders.module';
 import { APP_GUARD } from '@nestjs/core';
 import { TokenValidatorMiddleware } from 'src/core/auth/middleware/tokenValidator.middleware';
@@ -15,8 +8,13 @@ import { TopcoderModule } from 'src/shared/topcoder/topcoder.module';
 import { OriginRepository } from './repository/origin.repo';
 import { TaxFormRepository } from './repository/taxForm.repo';
 import { PaymentMethodRepository } from './repository/paymentMethod.repo';
+import { WinningsRepository } from './repository/winnings.repo';
 import { WebhooksModule } from './webhooks/webhooks.module';
 import { PaymentProvidersModule } from './payment-providers/payment-providers.module';
+import { AdminModule } from './admin/admin.module';
+import { WinningsModule } from './winnings/winnings.module';
+import { UserModule } from './user/user.module';
+import { WalletModule } from './wallet/wallet.module';
 
 @Module({
   imports: [
@@ -24,14 +22,12 @@ import { PaymentProvidersModule } from './payment-providers/payment-providers.mo
     TopcoderModule,
     PaymentProvidersModule,
     WebhooksModule,
+    AdminModule,
+    WinningsModule,
+    UserModule,
+    WalletModule,
   ],
-  controllers: [
-    HealthCheckController,
-    AdminWinningController,
-    UserWinningController,
-    WinningController,
-    WalletController,
-  ],
+  controllers: [HealthCheckController],
   providers: [
     {
       provide: APP_GUARD,
@@ -41,12 +37,10 @@ import { PaymentProvidersModule } from './payment-providers/payment-providers.mo
       provide: APP_GUARD,
       useClass: RolesGuard,
     },
-    AdminWinningService,
-    WinningService,
-    WalletService,
     OriginRepository,
     TaxFormRepository,
     PaymentMethodRepository,
+    WinningsRepository,
   ],
 })
 export class ApiModule implements NestModule {
