+++ src/api/repository/paymentMethod.repo.ts
@@ -1,8 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import {
-  PaymentMethodQueryResult,
-  UserPaymentMethodStatus,
-} from 'src/dto/paymentMethod.dto';
+import { payment_method_status } from '@prisma/client';
 import { PrismaService } from 'src/shared/global/prisma.service';
 
 @Injectable()
@@ -15,34 +12,15 @@ export class PaymentMethodRepository {
    * @param userId user id
    * @param tx transaction
    */
-  async hasVerifiedPaymentMethod(userId: string, tx?): Promise<boolean> {
-    const methods = await this.findPaymentMethodByUserId(userId, tx);
-    for (const method of methods) {
-      if (
-        method.status ===
-        UserPaymentMethodStatus.UserPaymentMethodStatusConnected.toString()
-      ) {
-        return true;
-      }
-    }
-    return false;
-  }
+  async hasVerifiedPaymentMethod(userId: string): Promise<boolean> {
+    const connectedUserPaymentMethod =
+      await this.prisma.user_payment_methods.findFirst({
+        where: {
+          user_id: userId,
+          status: payment_method_status.CONNECTED,
+        },
+      });
 
-  /**
-   * Get user payment methods
-   *
-   * @param userId user id
-   * @param tx transaction
-   * @returns payment methods
-   */
-  private async findPaymentMethodByUserId(userId: string, tx?) {
-    const db = tx || this.prisma;
-    const ret = await db.$queryRaw`
-      SELECT pm.payment_method_id, pm.payment_method_type, pm.name, pm.description, upm.status, upm.id
-      FROM payment_method pm
-      JOIN user_payment_methods upm ON pm.payment_method_id = upm.payment_method_id
-      WHERE upm.user_id=${userId}
-    `;
-    return (ret || []) as PaymentMethodQueryResult[];
+    return !!connectedUserPaymentMethod;
   }
 }
