+++ src/dto/date-filter.type.ts
@@ -1,0 +1,5 @@
+export enum DateFilterType {
+  LAST7DAYS = 'last7days',
+  LAST30DAYS = 'last30days',
+  ALL = 'all',
+}
