+++ src/dto/user.type.ts
@@ -1,0 +1,5 @@
+export class UserInfo {
+  id: string;
+  handle: string;
+  email: string;
+}
